Change event listeners to use defined methods
The previous way of just calling an instance method that matches the name of the event was too open to abuse - people could just call a "destructor" event and wreck havoc.

Now, a defined "listens" method will specify the events listened to, and the corresponding method name that will be called when the event is fired.