Fix infinite loop when impersonating users as a non-superadmin
Fixes #357. 

There was an infinite loop that occurred when attempting to impersonate users as a user that had the backend.impersonate_users permission but was not themselves a super user. This was caused because BackendAuth::getImpersonator() returned a separate user instance than the current user being evaluated despite the underlying DB row being the same.