Undefined variable $packagePath
When the package is not present in package.json, the correct error is not thrown because $packagePath not exists. I'm not sure what kind of feedback you would like to send, with this edit it'll be something like:

"Unable to watch "packagename", packagename was not found in the package.json's workspaces.packages property. Try running mix:install first."

Cheers!