Adjust early termination of requests from success/error
The "success" and "error" callbacks would previously stop any further request handling from taking place if the handler returned a falsy value, including nothing at all. This is inconsistent with the previous framework and with our docs.

This change makes it so that only an explicit "false" value returned from the "success" or "error" callback will stop the request handling there and then.