Revert "Fix default value for taglist formwidget in relation mode"
This caused an issue with taglists where existing relation values were being converted into their IDs instead of the actual nameFrom value. 

If Post belongsToMany Authors and the following configuration is used:
```yaml
authors:
    label: Authors
    type: taglist
    mode: relation
    separator: comma
    customTags: true
    nameFrom: full_name
```
Then loading a form with the authors relation already set would cause the ID of the authors to be displayed in the field instead of the full name and would cause issues when attempting to save the form.