Clone the form model before passing it to the RelationController
This prevents issues caused by extensions to the RelationController making changes to the relation's parent model for use cases needed in relations. These changes were previously also affecting the model instance used by the FormController behaviour.

If it is desired to interact with the exactly model instance as used by the FormController behaviour inside of your extensions to the RelationController you can use $controller->formGetModel() instead.