Allow sorting of lists by invisible columns
A previous bug fix applied here (https://github.com/wintercms/winter/commit/18af35d928956b3ebae636a0996add74d683bec8) forced only visible columns to be sortable to prevent an error from occurring if someone sorts by a column that is then removed from the columns list.

This change results in the same fix, but simply checks that the column is sortable (which would be false if the column no longer exists). This would  allow columns that are invisible but are also sortable to still be used for sorting purposes.

Fixes #796, #709
Refs:
- https://github.com/octobercms/october/issues/5227