Make into nullable type hint.
This is causing an error when using relation mode and you have an empty relation and try to save. the method expects an array, but is null so we either put the null value in an array or we can change the type to a nullable type hint. I'm fine with either and will change accordingly.