Convert Snowboard requests to classes
For some reason, these were not true importable classes as they were using the global Snowboard variable, so they couldn't be used in custom Snowboard builds.

I've created two new build files that use the same names as the previous builds, so this should be BC, but the source files are now true classes so they can be imported at will.