Add create:test command
Replaces the make:test command from Laravel.

Would be nice to add autocomplete support for the "test" argument to put in the path to any class in the plugin to automatically stub out the test based on information about that class and generate the test file in a matching relative path under the tests folder but one thing at a time for now.