Remove initial preload implementation.
The preload definition has changed somewhat since this was last looked at (see https://web.dev/articles/fetch-priority#history; ex. fetchpriority instead of priority).

Additionally, the MDN article on rel=preload (https://developer.mozilla.org/en-US/docs/Web/HTML/Attributes/rel/preload#the_basics) makes a good point that adding preloads for resources immediately before actually loading them doesn't actually help ("the browser probably discovers the <link rel="stylesheet"> and <script> elements in the same chunk of HTML as the preloads"), but should instead be used for resources loaded much later on in the page lifecycle:

> the benefits can be seen much more clearly the later resources are discovered and the larger they are. For example:
> Resources that are pointed to from inside CSS, like fonts or images.
> Resources that JavaScript can request, like imported scripts.

For this reason we'll remove the initial stuff we have related to fetchpriority and preloading. If we would like to implement it in the future, I would recommend that we take more of an approach of having a separate categorization for preloads, i.e. more along the lines of addPreload(). Potentially at that point we could do things like advance scanning of the assets added by addCss / addJs to dynamically generate some useful preloads, but for now this is not required by any active projects so we'll table it for now.