Add asset priority & use AssetMaker for global backend assets (#1022)
Co-authored-by: Luke Towers <luke@luketowers.ca>
Co-authored-by: Luke Towers <git@luketowers.ca>

This PR changes the global backend assets that were originally specified directly in the "head" template for the Backend, and uses the AssetMaker to load them instead. This should allow people the ability to manipulate the assets if need be through events.

To ensure that the global assets are loaded first before all others, I have also introduced a order value for assets. It may be specified by providing the second argument for addJs(), addCss(), etc. as an array and populating it with the following:

$this->addCss('my.css', [
    'build' => 'core',
    'order' => 50, // must be an integer
]);

The default order for all assets is 500.