Fix nested form data in Snowboard requests (#1182)
This fix allows nested objects within the request data to be correctly converted to nested form data when sending the request.

Previously, if the data option in a Request contained nested objects or arrays within the data, these were being incorrectly converted to [Array array] or [Object object], as the FormData object would simply convert them to strings. With this fix, we traverse the data structure and ensure that they are given the right key in order to keep the shape of the data in the POST data.