Change multiple conditions to use chained met elements
Since multiple conditions are considered "AND" conditions, we will chain the matched elements from the previous rule and only consider these in the following rule. This narrows the scope of successful rulesets and allows rules to work more congruently together.

For example, previously, if you used a "checked" condition with a "within" condition, if the "checked" condition passed, the "within" condition would pass if any target element (even those not checked) successfully passed the "within" condition. If you wanted the "within" condition to only consider the elements that were checked, it couldn't do so. Now that the chained elements are passed through, this will work.