Code quality fixes, discard repeated package registrations
In some cases, such as unit tests, a package might be registered twice with the same package name and config. Rather than erroring out, we'll just discard the repeated registration. Note that if a config file is registered twice with different package names, or two config files are registered to the same package name, this is still an error.