Fix issue caused by symfony polluting the environment
https://github.com/symfony/symfony/pull/53632 made it so that the --silent option is now provided & managed by the symfony core itself which would have been all well and good, except for the fact that since https://github.com/symfony/symfony/pull/24425 symfony has been persisting the selected verbosity layer in the environment. This means that if you have a test that calls a command with the --silent flag set, and then it's followed by a test that relies on the --silent flag not being set, that second test (and any future test that expects the --silent flag not to be set) will also fail.

This leads to all sorts of headaches around tests failing depending on the order they are run in, whether or not you're running a single test case or multiple, or even whether the environment you're running the tests in actually supports putenv().

Massive credit to @jaxwilko for finding the issue and coming up with the solution.